requirements.txt
# Healthcare AI Safety Guardrails System
# Production dependencies

# Core
openai>=1.0.0
anthropic>=0.18.0
numpy>=1.24.0
pyyaml>=6.0

# Optional: LangChain integration
# langchain>=0.1.0
# langchain-openai>=0.0.5

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0

# Type checking
mypy>=1.5.0
types-PyYAML>=6.0.0
